@sharkdp 2021-11-12
Thank you very much for your contribution. I initially restricted this on purpose, but I guess it's really helpful for debugging. This has been requested a few times.

Have you seen https://github.com/sharkdp/hyperfine/issues/276 and https://github.com/sharkdp/hyperfine/pull/282? I think it would be worth to go over a few discussion points before we merge this.

> The "± 0.00" output in "faster than" should probably be adjusted too,

Yes, I think so.