@avar 2021-11-14
> > The "± 0.00" output in "faster than" should probably be adjusted too,
> Yes, I think so.

Would you mind showing me how to do that? I tried that this morning and gave up on it. Surely something I can find out if I go through the full rust tutorial, but won't have time for that any time soon.

Basically I wanted to get the `max_runs` in `main.rs` down to `write_benchmark_comparison`, but it's not a `u64` like `min`, but in some option parser container type that I couldn't figure out how to unbox.