@sharkdp 2021-11-14
Done. I decided to make the standard-deviation fields `Optional`. This way,, we are forced to deal with missing values at every single place in the code base. All of the export formats also needed to be adapted:
- For Markdown, we just omit the standard deviation
- For CSV, we still use "0" as a default to not destroy the numerical format
- For JSON, we use `null` to indicate the missing value
- For AsciiDoc, we print `± ?` for now.